+++ modules/apps/foundation/portal/portal-upgrade/src/main/java/com/liferay/portal/upgrade/release/BaseUpgradeRelease.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.kernel.upgrade;
+package com.liferay.portal.upgrade.release;
 
 import com.liferay.portal.kernel.model.dao.ReleaseDAO;
+import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.StringBundler;
